+++ src/diagnostic/qbskeildiagnosticparser.ts
@@ -6,7 +6,7 @@ import * as QbsUtils from '../qbsutils';
 import {QbsDiagnosticParser} from './qbsdiagnosticutils';
 
 const ARM_CC_REGEXP = /"(.+\.\S+)",\sline\s(\d+):\s(Error|Warning):\s+(#\d+):\s(.+)$/;
-const ARM_CLANG_REGEXP = /^(.+\.\S+):(\d+):(\d+):\s(error|warning):\s(.+)/;
+const ARM_CLANG_REGEXP = /^(.+\.\S+):(\d+):(\d+):\s(fatal error|error|warning):\s(.+)/;
 const MCS_COMPILER_REGEXP = /^\*{3}\s(ERROR|WARNING)\s(.+)\sIN\sLINE\s(\d+)\sOF\s(.+\.\S+):\s(.+)$/;
 const MCS_ASSEMBLER_REGEXP = /^\*{3}\s(ERROR|WARNING)\s(.+)\sIN\s(\d+)\s\((.+),\sLINE\s\d+\):\s(.+)$/;
 
@@ -120,11 +120,15 @@ export class QbsKeilDiagnosticParser extends QbsDiagnosticParser {
     }
 
     private static encodeSeverity(severity: string): vscode.DiagnosticSeverity {
-        severity = severity.toLowerCase();
-        if (severity === 'error')
+        const s = severity.toLowerCase();
+        switch (s) {
+        case 'error':
+        case 'fatal error':
             return vscode.DiagnosticSeverity.Error;
-        else if (severity === 'warning')
+        case 'warning':
             return vscode.DiagnosticSeverity.Warning;
-        return vscode.DiagnosticSeverity.Hint;
+        default:
+            return vscode.DiagnosticSeverity.Information;
+        }
     }
 }