+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.0.5
 
 - Added the project name `${projectName}` substitution.
+- Fixed displaying for the 'note' and 'remark' messages in the `problems`
+pane with the 'information' severity level for GCC toolchains.
 
 ## 1.0.4
 - Default `qbs-configurations.json` file is not created unless user runs